+++ src/components/common/badges/PriorityBadge.js
@@ -1,4 +1,4 @@
-import Badge from '../Badge';
+import Badge from './Badge';
 
 const PriorityBadge = ({ priority, ...props }) => {
   const getVariant = (priority) => {
