+++ src/components/common/badges/StatusBadge.js
@@ -1,4 +1,4 @@
-import Badge from '../Badge';
+import Badge from './Badge';
 
 const StatusBadge = ({ status, ...props }) => {
   const getVariant = (status) => {
